movieBookingApp/src/app/forgot-username/forgot-username.component.spec.ts
import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
import { ReactiveFormsModule } from '@angular/forms';
import { AuthService } from '../auth.service';
import { Router } from '@angular/router';
import { of, throwError } from 'rxjs';
import { ForgotUsernameComponent } from './forgot-username.component';
import { RouterTestingModule } from '@angular/router/testing';

describe('ForgotUsernameComponent', () => {
  let component: ForgotUsernameComponent;
  let fixture: ComponentFixture<ForgotUsernameComponent>;
  let authService: jasmine.SpyObj<AuthService>;
  let router: jasmine.SpyObj<Router>;

  beforeEach(() => {
    const authServiceSpy = jasmine.createSpyObj('AuthService', ['forgotUserName']);
    const routerSpy = jasmine.createSpyObj('Router', ['navigate']);

    TestBed.configureTestingModule({
      imports: [ReactiveFormsModule, RouterTestingModule],
      declarations: [ForgotUsernameComponent],
      providers: [
        { provide: AuthService, useValue: authServiceSpy },
        { provide: Router, useValue: routerSpy },
      ],
    }).compileComponents();

    fixture = TestBed.createComponent(ForgotUsernameComponent);
    component = fixture.componentInstance;
    authService = TestBed.inject(AuthService) as jasmine.SpyObj<AuthService>;
    router = TestBed.inject(Router) as jasmine.SpyObj<Router>;

    router.navigate.and.stub();
  });

  beforeEach(() => {
    fixture.detectChanges();
  });

  it('should create the component', () => {
    expect(component).toBeTruthy();
  });

  it('should initialize the form correctly', () => {
    expect(component.forgotUserNameForm).toBeDefined();
    expect(component.forgotUserNameForm.get('email')).toBeDefined();
    expect(component.forgotUserNameForm.get('password')).toBeDefined();
  });

  it('should close the popup and navigate to login', () => {
    component.closePopup();
    expect(component.isPopupVisible).toBe(false);
    expect(router.navigate).toHaveBeenCalledWith(['/login']);
  });
});
